requirements.txt
Pillow~=11.2.1
pytest~=8.4.1
opencv-python~=4.11.0.86
matplotlib~=3.10.3
numpy~=2.1.3
torch~=2.7.1
torchvision~=0.22.1
streamlit~=1.47.0
pandas~=2.3.0